small update, micro optimization
